Optionally compute and save to the QC summary additional QC metrics.
Capturing QC metrics for an optional additional consensus file, which
is passed to the script by the new --fasta_amd argument, is added to
qc.py. Both the default and alternative minimum depths that are used
by ivar are also passed (new optional arguments --ivar_md and -ivar_amd)
and saved in the QC summary.